examples/image_to_stats/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
filelock==3.6.0
huggingface-hub==0.5.1
idna==3.3
joblib==1.1.0
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.53
six==1.16.0
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.19.1
typing_extensions==4.2.0
urllib3==1.26.9
